Linear conflict: add two to the Manhattan priority function whenever two tiles are in their goal row (or column) but are reversed relative to their goal position.
Pattern database: For each possible configuration of 4 tiles and the blank, determine the minimum number of moves to put just these tiles in their proper position and store these values in a database. The heuristic value is the maximum over all configurations, plus the number of moves made so far. This can reduce the number of search nodes examined for random 15-puzzle instances by a factor of 1000.